requirements.txt
flask
flask-login
flask-sqlalchemy
flask-migrate
Flask-WTF
flask-bootstrap4
opencv-python